# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hunar Screening Console API",
    "description": "Backend for two applications: AI voice screening of job applicants, and sourcing plus consent-gated voice outreach to prospects.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/hiring/jobs/extract": {
      "post": {
        "tags": [
          "hiring"
        ],
        "summary": "Extract From Description",
        "description": "Read a job description and fill in the whole create-role form.\n\nThe recruiter already has the description. Making them retype the\ntitle, the city and five questions it already contains is work the\ntext can do itself.\n\nSaves nothing. The result is a draft the recruiter reviews and edits,\nand `source` says whether a model read the description or whether it\nwas filled in by keyword matching, so nobody has to guess how much to\ntrust it.",
        "operationId": "extract_from_description_api_v1_hiring_jobs_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hiring/jobs/preview": {
      "post": {
        "tags": [
          "hiring"
        ],
        "summary": "Preview Agent",
        "description": "Render the agent script and extraction schema for a draft role.",
        "operationId": "preview_agent_api_v1_hiring_jobs_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hiring/jobs": {
      "get": {
        "tags": [
          "hiring"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_v1_hiring_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "type": "array",
                  "title": "Response List Jobs Api V1 Hiring Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "hiring"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_v1_hiring_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hiring/jobs/{job_id}": {
      "get": {
        "tags": [
          "hiring"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_hiring_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "hiring"
        ],
        "summary": "Update Job",
        "operationId": "update_job_api_v1_hiring_jobs__job_id__patch",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "hiring"
        ],
        "summary": "Archive Job",
        "description": "Archive rather than delete. Screening results describe real calls.",
        "operationId": "archive_job_api_v1_hiring_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentPreview": {
        "properties": {
          "introduction": {
            "type": "string",
            "title": "Introduction"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "agent_prompt": {
            "type": "string",
            "title": "Agent Prompt"
          },
          "result_prompt": {
            "type": "string",
            "title": "Result Prompt"
          },
          "result_schema": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Result Schema"
          },
          "variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Variables"
          }
        },
        "type": "object",
        "required": [
          "introduction",
          "objective",
          "agent_prompt",
          "result_prompt",
          "result_schema",
          "variables"
        ],
        "title": "AgentPreview",
        "description": "Exactly what the voice agent will be told, shown before any call.\n\nThis is the screen that makes the product legible: a recruiter can\nread the script and the extraction schema their job description\nproduced, and correct it, rather than discovering a bad prompt from a\nconfused candidate on the phone."
      },
      "AnswerType": {
        "type": "string",
        "enum": [
          "BOOLEAN",
          "NUMBER",
          "STRING",
          "ENUM"
        ],
        "title": "AnswerType",
        "description": "How an answer is coerced, rendered and scored.\n\nHunar returns every value as a string, so this is our declaration of\nwhat the string is meant to represent. It drives the normaliser, the\ncolumn renderer and the scoring rules from one place."
      },
...
      "DraftQuestion": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label",
            "description": "Short column heading, two or three words"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "The question as the agent should speak it"
          },
          "answer_type": {
            "type": "string",
            "enum": [
              "STRING",
              "NUMBER",
              "BOOLEAN",
              "ENUM"
            ],
            "title": "Answer Type"
          },
          "enum_options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enum Options",
            "description": "Allowed answers. Only for ENUM, otherwise empty."
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "description": "Importance from 0 to 5"
          },
          "is_knockout": {
            "type": "boolean",
            "title": "Is Knockout",
            "description": "True only for genuine hard requirements stated in the description"
          }
        },
        "type": "object",
        "required": [
          "label",
          "text",
          "answer_type",
          "weight",
          "is_knockout"
        ],
        "title": "DraftQuestion",
        "description": "One screening question the model proposed."
      },
      "ExtractRequest": {
        "properties": {
          "jd_text": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Jd Text"
          }
        },
        "type": "object",
        "required": [
          "jd_text"
        ],
        "title": "ExtractRequest",
        "description": "A pasted job description, to be turned into a filled-in form."
      },
      "FieldSpec": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "answer_type": {
            "$ref": "#/components/schemas/AnswerType"
          },
          "enum_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Options"
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "default": 0.0
          },
          "is_knockout": {
            "type": "boolean",
            "title": "Is Knockout",
            "default": false
          },
          "system": {
            "type": "boolean",
            "title": "System",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "answer_type"
        ],
        "title": "FieldSpec",
        "description": "One column of the results table, derived from a question.\n\nReturned alongside the rows so the frontend can build a table for a\nschema it has never seen. ``system`` marks the fields every job gets\nregardless of what the recruiter asked."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Title"
          },
          "company_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Company Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "description_raw": {
            "type": "string",
            "maxLength": 20000,
            "title": "Description Raw",
            "default": ""
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "ENGLISH"
          },
          "voice_persona": {
            "type": "string",
            "title": "Voice Persona",
            "default": "NEHA"
          },
          "persona_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Name"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Asia/Kolkata"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionInput"
            },
            "type": "array",
            "maxItems": 15,
            "minItems": 1,
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "title",
          "company_name",
          "questions"
        ],
        "title": "JobCreate"
      },
      "JobDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "voice_persona": {
            "type": "string",
            "title": "Voice Persona"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count",
            "default": 0
          },
          "call_count": {
            "type": "integer",
            "title": "Call Count",
            "default": 0
          },
          "completed_count": {
            "type": "integer",
            "title": "Completed Count",
            "default": 0
          },
          "shortlisted_count": {
            "type": "integer",
            "title": "Shortlisted Count",
            "default": 0
          },
          "in_flight_count": {
            "type": "integer",
            "title": "In Flight Count",
            "default": 0
          },
          "description_raw": {
            "type": "string",
            "title": "Description Raw"
          },
          "persona_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Name"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "hunar_agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hunar Agent Id"
          },
          "agent_synced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Synced At"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "type": "array",
            "title": "Questions"
          },
          "field_spec": {
            "items": {
              "$ref": "#/components/schemas/FieldSpec"
            },
            "type": "array",
            "title": "Field Spec"
          },
          "preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentPreview"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "company_name",
          "language",
          "voice_persona",
          "status",
          "created_at",
          "description_raw",
          "timezone"
        ],
        "title": "JobDetail"
      },
      "JobDraft": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "voice_persona": {
            "type": "string",
            "title": "Voice Persona"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/DraftQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "source": {
            "type": "string",
            "enum": [
              "model",
              "heuristic"
            ],
            "title": "Source",
            "default": "heuristic"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "title",
          "company_name",
          "location",
          "language",
          "voice_persona",
          "questions"
        ],
        "title": "JobDraft",
        "description": "Everything the create-role form needs, derived from the description."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "READY",
          "CALLING",
          "DONE",
          "ARCHIVED"
        ],
        "title": "JobStatus",
        "description": "Where a role is in its screening lifecycle."
      },
      "JobSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "voice_persona": {
            "type": "string",
            "title": "Voice Persona"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count",
            "default": 0
          },
          "call_count": {
            "type": "integer",
            "title": "Call Count",
            "default": 0
          },
          "completed_count": {
            "type": "integer",
            "title": "Completed Count",
            "default": 0
          },
          "shortlisted_count": {
            "type": "integer",
            "title": "Shortlisted Count",
            "default": 0
          },
          "in_flight_count": {
            "type": "integer",
            "title": "In Flight Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "company_name",
          "language",
          "voice_persona",
          "status",
          "created_at"
        ],
        "title": "JobSummary"
      },
      "JobUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "description_raw": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Raw"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "voice_persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Persona"
          },
          "persona_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Name"
          },
          "questions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/QuestionInput"
                },
                "type": "array",
                "maxItems": 15,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Questions"
          }
        },
        "type": "object",
        "title": "JobUpdate"
      },
...
      "PreviewRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title",
            "default": "the role"
          },
          "company_name": {
            "type": "string",
            "maxLength": 200,
            "title": "Company Name",
            "default": "the company"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "description_raw": {
            "type": "string",
            "maxLength": 20000,
            "title": "Description Raw",
            "default": ""
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "ENGLISH"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionInput"
            },
            "type": "array",
            "maxItems": 15,
            "minItems": 1,
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "questions"
        ],
        "title": "PreviewRequest",
        "description": "Ask what the agent would be told, without saving anything.\n\nPowers the live preview in the create-job form. Making the generated\nscript visible before any call is placed is what lets a recruiter\ncatch a bad prompt themselves, rather than hearing about it from a\nconfused candidate."
      },
      "QuestionInput": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Text"
          },
          "label": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Label"
          },
          "answer_type": {
            "$ref": "#/components/schemas/AnswerType",
            "default": "STRING"
          },
          "enum_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Options"
          },
          "extraction_hint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Hint"
          },
          "weight": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Weight",
            "default": 1.0
          },
          "is_knockout": {
            "type": "boolean",
            "title": "Is Knockout",
            "default": false
          },
          "scoring_rule": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoring Rule"
          }
        },
        "type": "object",
        "required": [
          "text",
          "label"
        ],
        "title": "QuestionInput",
        "description": "One question as the recruiter defines it."
      },
      "QuestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "field_key": {
            "type": "string",
            "title": "Field Key"
          },
          "answer_type": {
            "$ref": "#/components/schemas/AnswerType"
          },
          "enum_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Options"
          },
          "extraction_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Hint"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "is_knockout": {
            "type": "boolean",
            "title": "Is Knockout"
          },
          "scoring_rule": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoring Rule"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_index",
          "text",
          "label",
          "field_key",
          "answer_type",
          "weight",
          "is_knockout"
        ],
        "title": "QuestionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}